backend/pyproject.toml
[project]
name = "nutripilot-backend"
version = "0.1.0"
description = "NutriPilot AI - Autonomous Nutrition Agent Backend"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "NutriPilot Team" }
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "google-generativeai>=0.3.0",
    "opik>=1.0.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]
